MNT: move copy-copy function to utils module
